Improve rSrc's watch strategy to detect imgObj mutations
We not only need to deal with imgObj being replaced or appearing for
the first time, but also with it mutating. In that case, we need to be
watching an actual image address to notice the difference.

rSrc now watches on the object itself until it is defined for the first
time, and then watches on one of its properties so it can detect even
changes caused by object mutation.