Add a promise that keeps track of the chosen image's loading status
In some situations we need to know how the image's loading is going
from outside this directive. This is not something that can be
accomplished in a satisfying way because it's not trivial to determine
which image rSrc is going to load.
Instead, we can simply use JS to load the image and set it only after
we successfully loaded it, and expose a promise that keeps track of
this.