Improve error reporting from rSrc in case of non-undefined input
rSrc should be more aggressive to report errors when handed an object
that is neither undefined, nor a promise or a broken angular-resource
instance with $promise.
In that case, it's an invalid input just the same as if it was a
promise that resolved to such an object.
Undefined is not invalid input as we can not now for sure if it's not
just asynchonous input that has just not loaded yet.