During execution, module reports errors instead of throwing
We need to report errors and stop execution so that the behavior
remains consistent and debugging is facilitated. However, it is not
appropriate to completely stop by throwing when valid values can
be bound after invalid values and thus require the directive to keep
executing to update the relevant values afterward.