Make the image matcher more resilient in sub-optimal situations
Until now, the image matcher simply threw when it encountered a
situtation when it could not find the best image. This is good in
development as it helps catch programmer errors but it makes it less
robust in the face of unexpected situations. It now defaults to the
largest image in that case and only throws when there is literally no
image to be found.